feat(tracing): Allow trace collector to be unmeshed
The proxy currently treats the trace collector endpoint as a control plane component, mostly to prevent a circular dependency in the service stack. This means that we've had to manually supply a mesh identity in the form of a service account name, usually through an extension's injector.

However, this isn't strictly required as the control plane stack works just fine when the endpoint is unmeshed. This is realistically only useful for the trace collector which don't benefit much from being meshed, so this change only allows the collector to omit a mesh identity.

Signed-off-by: Scott Fleener <scott@buoyant.io>