nit(app/core): add banners to `impl` groups
we follow a convention in this codebase where `// === impl T ===`
comments dilineate groups of `impl` blocks for different types.

this style is particularly helpful when code is folded in a file,
because it helps clarify where traits for a given type are implemented.

this style is also helpful for grep'ing for interfaces, trait
implementations, etc., for a given type, since these may not always live
within a single file.

this commit adds some missing banner comments to our metric label types
in `linkerd_app_core`.

Signed-off-by: katelyn martin <kate@buoyant.io>