chore(app/core): remove `dead_code` allowance for `Stack<S>`
this `impl` block is ~350 lines, and contains the majority of our
`linkerd_app_core::svc::Stack<S>` interfaces.

this does not need to be marked with this `dead_code` allowance, because
these are public interfaces.

this commit removes this attribute. if future additions include unused
code, we can more narrowly affix this attribute to particular functions.

Signed-off-by: katelyn martin <kate@buoyant.io>