refactor(app/inbound): move `into_tcp_connect()` to `server`
`Inbound::into_tcp_connect()` is a method that takes the inbound proxy,
and prepares it for making TCP connections.

this is an implementation detail of the `Inbound::mk()` function, which
we call in the app's `Config::build()` startup code. it is not used by
any external callers.

this commit moves this method into the `server` submodule, next to the
four call-sites where it is invoked. the body of the function is
unchanged.

the `pub` directive is removed now that it lives alongside its dependent
code.

the doc comment for this function is fixed, which was previously a
fragment. the second line was a regular comment.

Signed-off-by: katelyn martin <kate@buoyant.io>