feat(app/inbound): count inbound requests (#4127)
this commit introduces a telemetry layer to the inbound proxy.

---

* refactor(app/inbound): add `linkerd-http-prom` dependency

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(app/inbound): `http::router` is `pub(crate)`

this will allow us to refer to metrics types, and invoke a `layer()`.

Signed-off-by: katelyn martin <kate@buoyant.io>

* feat(app/inbound): count inbound requests

this commit introduces a new `metrics::layer<N>()` function, and
installs it in our inbound http router.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(app/inbound): fully qualified `RequestCountFamilies`

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(app/inbound): introduce a `RequestCountFamilies` struct

this will provide us with a way to manage `inbound_http_request_total`
and `inbound_grpc_request_total` families based upon route
authorization.

Signed-off-by: katelyn martin <kate@buoyant.io>

* feat(app/inbound): count inbound `http` and `grpc` requests

Signed-off-by: katelyn martin <kate@buoyant.io>

* review(app/inbound): group `parent_` labels

we remove the `encode_label` closure, so that we do not have concurrent
mutable references to `enc`.

Signed-off-by: katelyn martin <kate@buoyant.io>

* nit(app/inbound): use `Permitted<T>` accessors

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>