refactor(http/prom): consistent middleware reëxports (#4151)
this pull request moves towards a consistent surface for the
`linkerd-http-prom` library.

```diff
 // linkerd/http/prom/src/lib.rs

 #![forbid(unsafe_code)]
 
 pub mod body_data;
-mod count_reqs;
+pub mod count_reqs;
 pub mod record_response;
-
-pub use self::count_reqs::{CountRequests, NewCountRequests, RequestCount, RequestCountFamilies};
```

this library contains a number of different telemetry middlewares for
counting requests, recording request and response durations, and for
measuring frames transmitted in request and response bodies.

the `count_reqs` types, unlike the other components in this library, are
no longer exported from the top-level of the library. this means that
call-sites in the inbound and outbound proxy have consistently shaped
imports for using out metrics layers.

Signed-off-by: katelyn martin <kate@buoyant.io>