refactor(app/inbound): metrics layer accepts `InboundMetrics` (#4166)
this commit makes a small refactor to the inbound proxy's
`metrics::layer()` middleware in the interest of future-proofing it
before additional metrics layers are introduced.

this will mean that the call site in `http/router.rs` will not need to
be updated repeatedly as we introduce request/response body frame size
metrics, duration histograms, and so forth.

Signed-off-by: katelyn martin <kate@buoyant.io>