refactor(app/inbound): consolidate route label encoding
in linkerd/linkerd2-proxy#4180, we noted during review that encoding
prometheus labels for route and parent metadata entails a slightly
unfortunate amount of repetitive boilerplate.

in the outbound proxy, we define a local `RouteLabels` type that
implements the `EncodeLabelSet` and `EncodeLabelSetMut`, which we use to
integrate label sets with the `prometheus_client` SDK.

this commit introduces an equivalent pattern to the inbound proxy. a
`RouteLabels` newtype wrapper over the shared `linkerd_app_core`
structure is introduced, which in turn simplifies the `EncodeLabelSet`
and `EncodeLabelSetMut` implementations of the `RequestCountLabels`,
`ResponseBodyDataLabels`, and `RequestBodyDataLabels` types.

Signed-off-by: katelyn martin <kate@buoyant.io>