refactor(http/prom): rename `NewRecordBodyData::layer_via()`
in linkerd/linkerd2-proxy#4174, we refactored this middleware so that it
is now agnostic to the particular metrics structure used by the
extractor, meaning that the constructor no longer accepts a
`RequestBodyFamilies<L>` parameter.

in the wake of this change, our request body frame size metrics
middleware has a constructor with a type signature like `X -> impl
Layer`.

elsewhere throughout the project, constructors that accept an extractor
and return a `svc::Layer` are typically called `layer_via()`.

this commit renames `NewRecordBodyData::new()` to
`NewRecordBodyData::layer_via()`.

Signed-off-by: katelyn martin <kate@buoyant.io>