refactor(app/outbound): use `svc::layers()` to apply layers
we apply a collection of `Layers` to an inner service in our metrics
subsystem. doing so with `layer::mk()` and `Layer::layer()` is somewhat
ungainly.

this commit relies on `svc::layers()` to apply a sequence of telemetry
layers in sequence.

this should have no mechanical difference with the preëxisting code.

Signed-off-by: katelyn martin <kate@buoyant.io>