fix(tracing): Flush spans on interval instead of idle timeout
Our trace span exporter uses batching logic to avoid making excessive requests to the trace collector. It does this by either waiting for the batch buffer to hit a maximum number of spans (currently 1000), or waiting for no spans to be collected in some timeout window (currently 10s).

If a workload emits spans at a slow but consistent rate, say 1 span/s, it would cause the timeout window to never be hit while the buffer slowly fills. At this rate, it would cause spans to not be emitted for a long time.

This replaces the timeout window with an interval timer that attempts to send the currently buffered spans on a consistent interval, regardless of if the collector has sat idle. This interval is set to 10s to match the current idle timeout, but this can be tuned in the future along with the buffer size if we feel the need.

Signed-off-by: Scott Fleener <scott@buoyant.io>