chore(tracing)!: Remove OpenCensus trace protocol
OpenCensus as a protocol has been sunset for some time now. Additionally, we have supported its replacement, OpenTelemetry, for a long time now, and we don't even expose the OpenCensus configuration from the linkerd control plane chart.

This removes OpenCensus from the proxy. There's some cleanup that will have to be done in follow-ups, but this attempts to keep the diff as minimal as possible.

Signed-off-by: Scott Fleener <scott@buoyant.io>