feat(http/body-eos): introduce `BodyWithEosFn<B, F>`
this commit introduces a piece of `http_body::Body` middleware.

`BodyWithEosFn<B, F>` wraps an inner `B`-typed body, and invokes an
`F`-typed callback when the end of the stream is reached.

detecting the end of a request body or response body stream is a tricky,
reoccuring problem in different pieces of our `tower` middleware.
streams can end with a `None`, a `Some(Ok(_))` trailers frame, a
`Some(Err(_))` error, can be dropped prior to completion, or a
`Some(Ok(_))` data frame that leaves the body reporting a `true` hint
from `Body::is_end_stream()`.

this is very easy to do incorrectly.

this piece of body middleware abstracts over all of these state
transitions, and allows callers to instead provide a callback that will
be invoked when the stream finishes.

Signed-off-by: katelyn martin <kate@buoyant.io>