feat(http/prom): add http and grpc status labelers
this commit introduces a collection of `MkStreamLabel` and `StreamLabel`
implementations that can be used to inspect status codes for HTTP and
gRPC traffic.

these do not emit duration labels, but can be used by other
implementors of these traits to perform the common logic of checking
HTTP status codes in response front matter, and for inspecting body
trailers in gRPC traffic.

Signed-off-by: katelyn martin <kate@buoyant.io>