feat(http/prom): introduce `StreamLabel` components (#4251)
see linkerd/linkerd2-proxy#4250 (59f90ac3) for previous related work.

* feat(http/prom): add http and grpc status labelers

this commit introduces a collection of `MkStreamLabel` and `StreamLabel`
implementations that can be used to inspect status codes for HTTP and
gRPC traffic.

these do not emit duration labels, but can be used by other
implementors of these traits to perform the common logic of checking
HTTP status codes in response front matter, and for inspecting body
trailers in gRPC traffic.

Signed-off-by: katelyn martin <kate@buoyant.io>

* feat(http/prom): add an error labeler

this commit introduces a `LabelError<E>` type, which provides a
`StreamLabel` implementation that maps boxed errors to labels.

this is generic across `E`-typed labels that can be constructed `From` a
reference to a boxed `linkerd_error::Error`.

Signed-off-by: katelyn martin <kate@buoyant.io>

* feat(http/prom): add an identity labeler

this commit introduces a simple "identity" implementation of
`MkStreamLabel` and `StreamLabel`.

this is useful for metrics such as our duration recording histogram,
that do not inspect the request, response, or outcome for further
labels.

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>