chore(tracing): Use upstream tracing batch exporter
The proxy used its own custom trace exporter for some time. It was built before the upstream OpenTelemetry libraries were available, and they have outlived their usefulness.

This replaces the custom exporter with a batch exporter configured to export to the same endpoint with the same service configuration.

In the future, this exporter can be installed as a global default trace processor, which would decouple it from the service layer where the proxy generates spans for requests.

Signed-off-by: Scott Fleener <scott@buoyant.io>