chore(tracing): Use upstream tracing propagation
The proxy used its own custom trace propagation for some time. It was built before the upstream OpenTelemetry libraries were available.

This replaces the manual propagation with a combination of the upstream w3c and b3 propagators, modified slightly to only propagate in one format or the other (and not both).

Signed-off-by: Scott Fleener <scott@buoyant.io>