feat(http/prom): add status counting middleware
this commit introduces a new submodule to `linkerd-http-prom`.

this submodule includes a `NewRecordStatusCode<N>` middleware that can
be used with the status code labeling implementations in `stream_label`
to record Prometheus metrics counting response status codes for HTTP and
gRPC traffic.

a test suite demonstrates that this works properly for HTTP and gRPC
traffic, including edge cases like errors or cancelled bodies.

Signed-off-by: katelyn martin <kate@buoyant.io>