feat(http/body-eos): `EosRef<'a, E>: Clone + Copy`
this commit introduces clone and copy implementations for the
`EosRef<'a, E>` type used to inspect and classify the outcome of a
request or response body.

as the documentation comment notes in the commit below, this enum only
contains immutable references to other values: a header map or an
error. other variants are empty tags.

```rust
pub enum EosRef<'a, E = Error> {
    None,
    Trailers(&'a HeaderMap),
    Error(&'a E),
    Cancelled,
}
```

this commit makes this type implicitly clonable.

Signed-off-by: katelyn martin <kate@buoyant.io>