feat(http/prom): introduce `NewRecordStatusCode` middleware (#4298)
* refactor(http/prom): status code labelers are `Clone`

these types do not contain any internal state, and could thus be
clonable. this commit makes `MkLabelGrpcStatus` and
`MkLabelHttpStatus` clonable.

Signed-off-by: katelyn martin <kate@buoyant.io>

* refactor(http/prom): constructable `RequestCancelled`

this commit removes the uninhabited `()` state in `RequestCancelled`.

this will allow our forthcoming status middleware to construct this
type when a request is cancelled.

Signed-off-by: katelyn martin <kate@buoyant.io>

* feat(http/prom): add status counting middleware

this commit introduces a new submodule to `linkerd-http-prom`.

this submodule includes a `NewRecordStatusCode<N>` middleware that can
be used with the status code labeling implementations in `stream_label`
to record Prometheus metrics counting response status codes for HTTP and
gRPC traffic.

a test suite demonstrates that this works properly for HTTP and gRPC
traffic, including edge cases like errors or cancelled bodies.

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>