fix(spiffe-proto): generate proto bindings without embedding package definition (#4324)
Our workload.proto package has a package definition in it. We rely on
that package definition for naming our generated protobuf file. We
however do not want the package details being part of the
service/method names of the generated client code.

For that purpose we have used the `disable_package_emission` method
successfully in the past in order to generate code that does not have the
package details in the service/method names.

When we moved to tonic_prost_build however in [this PR](https://github.com/linkerd/linkerd2-proxy/commit/bdf10e47#diff-4a9729833e19e5f140e358c372e53f7ccf5e23de7c23062e4282127512dd621f), we have lost this functionality as the `emit_package` method does not seem to be honored.

In order to fix that and revert back to generating code with the correct service definitions
while still preserving the name of the generated rust file, this PR:

  * removes the package definition from the protobuf file
  * adds a post processing step that renames the generated `_.rs` file to what it is supposed to be

---

* fix(spiffe-proto): generate proto bindings without embedding package definition

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>

* fmt

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>

* set default package name instead of renaming file

---------

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>
Co-authored-by: Alejandro Pedraza <alejandro@buoyant.io>