refactor(Dockerfile): use `ghcr.io/linkerd/proxy` base image
see linkerd/linkerd-proxy#4333 for previous context.

this commit makes changes to the Dockerfile provided in this repository,
for use in the proxy's development process.

rather than using `debian:bookworm-slim` as the base image, this commit
helps deduplicate the tricky business of setting networking capabilities
on executables needed when running as an init container.

this has one negative consequence, which is that we can no longer attach
to a `bash` shell in a running pod when using this image. this is
unfortunate, but in my experience isn't often needed by proxy
developers.

i believe that, should we need to revisit the need for a shell in this
image, we should do instead make use of the `Dockerfile-debug` image
provided in the linkerd2 repo.

if we ran a command like
`just docker --build-arg LINKERD2_IMAGE='ghcr.io/linkerd/debug:edge-25-11.3'`
we could specify the debug image as a base image instead, providing
developers not only with a shell, but also other helpful utilities like
`curl`, `tcpdump`, and so on.

unfortunately, this does not work today, because the image appears to no
longer be published, and has drifted from our latest edge release. i
have not pulled on that string further at the time of writing.

one explicit _benefit_ of the changes in this commit is that we bring
proxy development closer to the real world, meaning that CI in this
repository runs using the same image that the proxy will run inside of
in the linkerd2 repository and in typical clusters.

---

* linkerd/linkerd2#14348
* linkerd/linkerd2#14577
* linkerd/linkerd-proxy#4333

Signed-off-by: katelyn martin <kate@buoyant.io>