fix: proxy-init and policy-container images have been removed (#4333)
in linkerd/linkerd2#14577, we combined the proxy and proxy-init image.
in https://github.com/linkerd/linkerd2#14348, we removed the
policy-controller image. this branch aims to address these breaking
changes, and restore CI builds in this repository.

besides just combining images, \#14577 included some breaking changes to
the helm charts, specifically in how we configure the control plane.
this has affected us, because we do _not_ use the production-oriented
`ghcr.io/linkerd/proxy` image when running the `linkerd-install` test,
or running a local cluster via the commands provided in the `justfile`.

we use a development-oriented image provided
[here](https://github.com/linkerd/linkerd2-proxy/blob/main/Dockerfile)
so that we can attach a shell to the proxy container during development,
and run various networking utilities to e.g. diagnose bugs.

this caused us some problems because \#14577 removed the helm
configuration surface used to configure an image for the proxy-init
container. thus, pods would run the development image from the proxy
repo, rather than the `'ghcr.io/linkerd/proxy'` image that i've
attempted to specify in the justfile.

```
Init Containers:                                                                                                                                                
  linkerd-init:                                                                                                                                                 
    Container ID:    containerd://bfe539d4686401676687f3526c43611b90bf794c300c0c9797582a57ce13196d                                                              
    Image:           localhost/linkerd/proxy:kate.init-image-is-combined.c374b8d8c                                                                              
    Image ID:        sha256:e6bd21b9afa12fdac81f8b3df6c3ba035015a162165afbbba5cc26f729a49f9f                                                                    
    Port:            <none>                                                                                                                                     
    Host Port:       <none>                                                                                                                                     
    SeccompProfile:  RuntimeDefault                                                                                                                             
    Command:                                                                                                                                                    
      /usr/lib/linkerd/linkerd2-proxy-init
```

there are two ways to remediate this: (1) copy the same changes
involving `setcap` commands over to that development dockerfile (_i also
found that this required setting the user to `root` for things to
work_), or (2) revert parts of \#14577 and reintroduce these settings to
the control plane chart, instead defaulting to the
`cr.l5d.io/linkerd/proxy` image.

my understanding is that (2) would be unfortunate because we are
actively trying to get away from having to maintain too much
init-related configuration surface, as it introduces a significant
maintenance burden in the control plane.

this branch updates the no-longer-recent edge release used in our
development image, and then updates the images used in our justfile.

changes from https://github.com/linkerd/linkerd2/pull/14577 are
additionally introduced here, so that the proxy image also functions as
a workable init container image.

---

* chore(Dockerfile): update `LINKERD2_IMAGE` edge version

Signed-off-by: katelyn martin <kate@buoyant.io>

* fix: address upstream changes to proxy-init image

Signed-off-by: katelyn martin <kate@buoyant.io>

---------

Signed-off-by: katelyn martin <kate@buoyant.io>